Moved scripts to geometry folder & namespace
