+++ WorldBuilder/Scripts/Editor/Geometry.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8cd9ba6315669a744b1a0de308717343
+guid: 7b604f2686d11ca47aadeb339956d685
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
