+++ WorldBuilder/Scripts/Editor/Geometry/Types.cs
